Arduino implementation of the Word Clock.
The split into an .h and an .ino file is mostly motivated by the Arduino IDE, which takes it upon itself to combine all C++ source code files into a single one and then transmogrify that combined source for, I guess, the programmer's convenience (by doing things like pre-declaring everything the IDE thinks is a function definition), which falls flat in the presence of such unorthodox things as having a custom type somewhere.

Anyway: The .h file contains most of the actual, modular, generic Word Clock logic, and the .ino file uses it to set up the specific Word Clock we want to make.