Added support for a DS3231 RTC module used in addition to the DCF77 receiver.
It turns out that this Arduino board at least has a surprising amount of drift in its own timekeeping; something in the order of 5-10%, much as if its crystal was running at something more like 17 MHz than the expected 16 MHz.

The RTC module is used to synchronize the internal timekeeping once every minute at :45 seconds, which is timed to have the least possible effect on the visible time transitions as well as the ten-second countdowns just prior to New Year's Eve and Vivi's birthday.

If and when the DCF77 decoder receives a viable timestamp, it updates both the RTC and the internal timekeeping.